third_party/WebKit/LayoutTests/battery-status/api-defined-expected.txt

Test basic API definitions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS promise is defined.
PASS promise.then is defined.
batteryStatusSuccess invoked
PASS battery is defined.
PASS battery is non-null.
PASS battery.charging is defined.
PASS battery.chargingTime is defined.
PASS battery.dischargingTime is defined.
PASS battery.level is defined.
PASS battery.charging is mockBatteryInfo.charging
PASS battery.chargingTime is mockBatteryInfo.chargingTime
PASS battery.dischargingTime is mockBatteryInfo.dischargingTime
PASS battery.level is mockBatteryInfo.level
PASS typeof battery.onchargingchange == 'object' is true
PASS typeof battery.onchargingtimechange == 'object' is true
PASS typeof battery.ondischargingtimechange == 'object' is true
PASS typeof battery.onlevelchange == 'object' is true
PASS 'onchargingchange' in battery is true
PASS 'onchargingtimechange' in battery is true
PASS 'ondischargingtimechange' in battery is true
PASS 'onlevelchange' in battery is true
PASS successfullyParsed is true

TEST COMPLETE
